requirements.txt
accelerate==0.34.0
aiohttp==3.9.1
aiosignal==1.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.71.0
anyio==4.9.0
apache-tvm-ffi==0.1.6
appdirs==1.4.4
astor==0.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==25.4.0
bitsandbytes==0.44.0
black==23.12.1
blake3==1.0.8
Brotli==1.1.0
cachetools==6.2.4
cbor2==5.7.1
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
coloredlogs==15.0.1
compressed-tensors==0.12.2
contourpy==1.3.2
cryptography==46.0.3
cuda-bindings==13.1.1
cuda-pathfinder==1.3.3
cuda-python==13.1.1
cupy-cuda12x==13.6.0
cycler==0.12.1
datasets==2.16.1
depyf==0.20.0
dill==0.3.7
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
einops==0.8.1
email-validator==2.3.0
exceptiongroup==1.2.2
executing==2.0.1
faiss-cpu==1.11.0
fastapi==0.127.0
fastapi-cli==0.0.20
fastapi-cloud-cli==0.8.0
fastar==0.8.0
fastrlock==0.8.3
filelock==3.20.1
fire==0.5.0
flash_attn==2.8.3
flashinfer-python==0.5.3
fonttools==4.61.1
frozenlist==1.4.1
fsspec==2023.10.0
gguf==0.10.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.10
ijson==3.4.0.post0
importlib_metadata==8.7.1
inflate64==1.0.0
interegular==0.3.3
ipython==8.19.0
jedi==0.19.1
Jinja2==3.1.6
jiter==0.12.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
langchain-text-splitters==0.3.8
lark==1.2.2
llguidance==1.3.0
llvmlite==0.44.0
lm-format-enforcer==0.10.6
loguru==0.7.3
loralib==0.1.2
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.8
matplotlib-inline==0.1.6
mcp==1.25.0
mdurl==0.1.2
mistral_common==1.8.8
mkl_fft==2.1.1
mkl_random==1.3.0
mkl-service==2.5.2
model-hosting-container-standards==0.1.12
mpmath==1.3.0
msgpack==1.1.2
msgspec==0.20.0
multidict==6.0.4
multiprocess==0.70.15
multivolumefile==0.2.3
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.13.0
numba==0.61.2
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cudnn-frontend==1.17.0
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu12==0.7.1
nvidia-cutlass-dsl==4.3.4
nvidia-ml-py==13.590.44
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.1.105
openai==2.8.1
openai-harmony==0.0.8
opencv-python-headless==4.11.0.86
optimum==1.16.1
orjson==3.10.17
outlines==0.0.46
outlines_core==0.2.11
packaging==25.0
pandas==2.3.3
parso==0.8.3
partial-json-parser==0.2.1.1.post7
pathspec==0.12.1
pillow==12.0.0
pip==25.3
prometheus_client==0.23.1
prometheus-fastapi-instrumentator==7.1.0
prompt-toolkit==3.0.43
protobuf==6.33.2
psutil==7.2.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
py7zr==0.20.8
pyairports==0.0.1
pyarrow==14.0.2
pyarrow-hotfix==0.6
pybase64==1.4.3
pybcj==1.0.2
pycountry==24.6.1
pycparser==2.23
pycryptodomex==3.19.1
pydantic==2.12.5
pydantic_core==2.41.5
pydantic-extra-types==2.10.6
pydantic-settings==2.12.0
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.3.1
pyppmd==1.1.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-multipart==0.0.21
pytorch-triton==2.2.0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
pyzstd==0.15.9
ray==2.53.0
referencing==0.37.0
regex==2023.12.25
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rich-toolkit==0.17.1
rignore==0.7.6
rpds-py==0.30.0
safetensors==0.7.0
scipy==1.15.3
seaborn==0.13.2
sentencepiece==0.2.0
sentry-sdk==2.48.0
setproctitle==1.3.7
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sse-starlette==3.0.4
stack-data==0.6.3
starlette==0.50.0
supervisor==4.3.0
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
texttable==1.7.0
tiktoken==0.9.0
tokenize-rt==5.2.0
tokenizers==0.22.1
tomli==2.0.1
torch==2.4.0
torchaudio==2.9.0
torchvision==0.19.0
tqdm==4.67.1
traitlets==5.14.1
transformers==4.57.3
triton==3.0.0
typer==0.20.1
typer-slim==0.20.1
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.40.0
uvloop==0.22.1
vllm==0.6.3
watchfiles==1.1.1
wcwidth==0.2.12
websockets==15.0.1
wheel==0.45.1
xformers==0.0.27.post2
xgrammar==0.1.27
xxhash==3.4.1
yarl==1.9.4
zipp==3.23.0
zstandard==0.23.0
